% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grangesUtils.R
\name{BSgenomeToDNAStringSet}
\alias{BSgenomeToDNAStringSet}
\title{Convert BSgenome to DNAStringSet}
\usage{
BSgenomeToDNAStringSet(BSgnm)
}
\arguments{
\item{BSgnm}{A BSgenome object}
}
\value{
A DNAStringSet object
}
\description{
Convert BSgenome to DNAStringSet
}
\examples{
BSgenomeToDNAStringSet(BSgenome.Celegans.UCSC.ce11::Celegans)
}
