% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grangesUtils.R
\name{BSgenomeToGR}
\alias{BSgenomeToGR}
\title{Convert BSgenome to GenomicRanges}
\usage{
BSgenomeToGR(BSgnm, defaultStrand = "*")
}
\arguments{
\item{BSgnm}{A BSgenome object}

\item{defaultStrand}{One of "+", "-" or "*" for the strand from which the sequences come from (default="*")}
}
\value{
A GenomicRanges object
}
\description{
Convert BSgenome to GenomicRanges
}
\examples{
BSgenomeToGR(BSgenome.Celegans.UCSC.ce11::Celegans)
}
