% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grangesUtils.R
\name{ucscToWbGR}
\alias{ucscToWbGR}
\title{Convert UCSC to wormbase chromosome names}
\usage{
ucscToWbGR(ucscGR)
}
\arguments{
\item{ucscGR}{A GRanges object with ucsc chomosome names ("chrI"..."chrM")}
}
\value{
A GenomicRanges object
}
\description{
Convert UCSC to wormbase chromosome names
}
\examples{
ucscToWbGR(BSgenomeToGR(BSgenome.Celegans.UCSC.ce11::Celegans))
}
