% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grangesUtils.R
\name{wbToUcscGR}
\alias{wbToUcscGR}
\title{Convert wormbase to ucsc chromosome names}
\usage{
wbToUcscGR(wbGR)
}
\arguments{
\item{wbGR}{A GRanges object with ucsc chomosome names ("chrI"..."chrM")}
}
\value{
A GenomicRanges object
}
\description{
Convert wormbase to ucsc chromosome names
}
\examples{
wbToUcscGR(ucscToWbGR(BSgenomeToGR(BSgenome.Celegans.UCSC.ce11::Celegans)))
}
